+++ src/alignment/std_traits.rs
@@ -4,9 +4,13 @@ use std::{
     simd::{LaneCount, SupportedLaneCount},
 };
 
-impl<const N: usize, const S: usize> PartialEq for LocalProfiles<'_, N, S>
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> PartialEq
+    for LocalProfiles<'_, M, N, O, P, S>
 where
+    LaneCount<M>: SupportedLaneCount,
     LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
 {
     fn eq(&self, other: &Self) -> bool {
         (self.query == other.query)
@@ -16,11 +20,22 @@ where
     }
 }
 
-impl<const N: usize, const S: usize> Eq for LocalProfiles<'_, N, S> where LaneCount<N>: SupportedLaneCount {}
-
-impl<const N: usize, const S: usize> Hash for LocalProfiles<'_, N, S>
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> Eq for LocalProfiles<'_, M, N, O, P, S>
 where
+    LaneCount<M>: SupportedLaneCount,
     LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
+{
+}
+
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> Hash
+    for LocalProfiles<'_, M, N, O, P, S>
+where
+    LaneCount<M>: SupportedLaneCount,
+    LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
 {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.query.hash(state);
@@ -30,9 +45,13 @@ where
     }
 }
 
-impl<const N: usize, const S: usize> PartialEq for SharedProfiles<'_, N, S>
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> PartialEq
+    for SharedProfiles<'_, M, N, O, P, S>
 where
+    LaneCount<M>: SupportedLaneCount,
     LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
 {
     fn eq(&self, other: &Self) -> bool {
         (self.query == other.query)
@@ -42,11 +61,22 @@ where
     }
 }
 
-impl<const N: usize, const S: usize> Eq for SharedProfiles<'_, N, S> where LaneCount<N>: SupportedLaneCount {}
-
-impl<const N: usize, const S: usize> Hash for SharedProfiles<'_, N, S>
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> Eq for SharedProfiles<'_, M, N, O, P, S>
 where
+    LaneCount<M>: SupportedLaneCount,
     LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
+{
+}
+
+impl<const M: usize, const N: usize, const O: usize, const P: usize, const S: usize> Hash
+    for SharedProfiles<'_, M, N, O, P, S>
+where
+    LaneCount<M>: SupportedLaneCount,
+    LaneCount<N>: SupportedLaneCount,
+    LaneCount<O>: SupportedLaneCount,
+    LaneCount<P>: SupportedLaneCount,
 {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.query.hash(state);